错误码处理完成

+++ app/model/base/reponse.go
@@ -1,7 +1,7 @@
 package base
 
 type Response struct {
-	ErrorCode    int64       `json:"error_code"`
+	ErrorCode    int         `json:"error_code"`
 	ErrorMessage string      `json:"error_message"`
 	Succeed      bool        `json:"succeed"`
 	Data         interface{} `json:"data"`
